refactor(settings): déplacer la section indexation de Connaissances vers Avancé
